rm cast_op

+++ cpp/include/cuvs/cluster/kmeans.hpp
@@ -220,7 +220,7 @@ void fit(raft::resources const& handle,
          std::optional<raft::host_vector_view<const float, int64_t>> sample_weight,
          raft::device_matrix_view<float, int64_t> centroids,
          raft::host_scalar_view<float> inertia,
-         raft::host_scalar_view<int64_t> n_iter);
+         raft::host_scalar_view<int> n_iter);
 
 /**
  * @brief Find clusters with k-means algorithm using batched processing of host data.
@@ -231,7 +231,7 @@ void fit(raft::resources const& handle,
          std::optional<raft::host_vector_view<const double, int64_t>> sample_weight,
          raft::device_matrix_view<double, int64_t> centroids,
          raft::host_scalar_view<double> inertia,
-         raft::host_scalar_view<int64_t> n_iter);
+         raft::host_scalar_view<int> n_iter);
 
 /**
  * @brief Find clusters with k-means algorithm.
@@ -282,7 +282,7 @@ void fit(raft::resources const& handle,
          std::optional<raft::device_vector_view<const float, int64_t>> sample_weight,
          raft::device_matrix_view<float, int64_t> centroids,
          raft::host_scalar_view<float> inertia,
-         raft::host_scalar_view<int64_t> n_iter);
+         raft::host_scalar_view<int> n_iter);
 
 /**
  * @brief Find clusters with k-means algorithm.
@@ -333,7 +333,7 @@ void fit(raft::resources const& handle,
          std::optional<raft::device_vector_view<const double, int64_t>> sample_weight,
          raft::device_matrix_view<double, int64_t> centroids,
          raft::host_scalar_view<double> inertia,
-         raft::host_scalar_view<int64_t> n_iter);
+         raft::host_scalar_view<int> n_iter);
 
 /**
  * @brief Find clusters with k-means algorithm.
@@ -888,7 +888,7 @@ void fit_predict(raft::resources const& handle,
                  std::optional<raft::device_matrix_view<float, int64_t>> centroids,
                  raft::device_vector_view<uint32_t, int64_t> labels,
                  raft::host_scalar_view<float> inertia,
-                 raft::host_scalar_view<int64_t> n_iter);
+                 raft::host_scalar_view<int> n_iter);
 
 /**
  * @brief Compute k-means clustering and predicts cluster index for each sample
@@ -943,7 +943,7 @@ void fit_predict(raft::resources const& handle,
                  std::optional<raft::device_matrix_view<double, int64_t>> centroids,
                  raft::device_vector_view<uint32_t, int64_t> labels,
                  raft::host_scalar_view<double> inertia,
-                 raft::host_scalar_view<int64_t> n_iter);
+                 raft::host_scalar_view<int> n_iter);
 
 /**
  * @brief Compute balanced k-means clustering and predicts cluster index for each sample
